+++ tests/e2e_tests/test_neuron_certificate.py
@@ -11,6 +11,7 @@ from tests.e2e_tests.utils.e2e_test_utils import (
 )
 
 
+@pytest.mark.skip(reason="Have to be rewritten from scratch")
 @pytest.mark.asyncio
 async def test_neuron_certificate(local_chain):
     """