+++ tests/unit_tests/test_subtensor.py
@@ -771,8 +771,8 @@ def test_get_total_subnets_no_block(mocker, subtensor):
 
 
 # `get_subnets` tests
-def test_get_subnets_success(mocker, subtensor):
-    """Test get_subnets returns correct list when subnet information is found."""
+def test_get_netuids_success(mocker, subtensor):
+    """Test get_netuids returns correct list when subnet information is found."""
     # Prep
     block = 123
     mock_netuid1 = mocker.MagicMock(value=1)
@@ -782,14 +782,14 @@ def test_get_subnets_success(mocker, subtensor):
     mocker.patch.object(subtensor, "query_map_subtensor", return_value=mock_result)
 
     # Call
-    result = subtensor.get_subnets(block)
+    result = subtensor.get_netuids(block)
 
     # Asserts
     assert result == [1, 2]
     subtensor.query_map_subtensor.assert_called_once_with("NetworksAdded", block)
 
 
-def test_get_subnets_no_data(mocker, subtensor):
+def test_get_netuids_no_data(mocker, subtensor):
     """Test get_subnets returns empty list when no subnet information is found."""
     # Prep
     block = 123
@@ -798,15 +798,15 @@ def test_get_subnets_no_data(mocker, subtensor):
     mocker.patch.object(subtensor, "query_map_subtensor", return_value=mock_result)
 
     # Call
-    result = subtensor.get_subnets(block)
+    result = subtensor.get_netuids(block)
 
     # Asserts
     assert result == []
     subtensor.query_map_subtensor.assert_called_once_with("NetworksAdded", block)
 
 
-def test_get_subnets_no_records_attribute(mocker, subtensor):
-    """Test get_subnets returns empty list when result has no records attribute."""
+def test_get_netuids_no_records_attribute(mocker, subtensor):
+    """Test get_netuids returns empty list when result has no records attribute."""
     # Prep
     block = 123
     mock_result = mocker.MagicMock()
@@ -814,15 +814,15 @@ def test_get_subnets_no_records_attribute(mocker, subtensor):
     mocker.patch.object(subtensor, "query_map_subtensor", return_value=mock_result)
 
     # Call
-    result = subtensor.get_subnets(block)
+    result = subtensor.get_netuids(block)
 
     # Asserts
     assert result == []
     subtensor.query_map_subtensor.assert_called_once_with("NetworksAdded", block)
 
 
-def test_get_subnets_no_block_specified(mocker, subtensor):
-    """Test get_subnets with no block specified."""
+def test_get_netuids_no_block_specified(mocker, subtensor):
+    """Test get_netuids with no block specified."""
     # Prep
     mock_netuid1 = mocker.MagicMock(value=1)
     mock_netuid2 = mocker.MagicMock(value=2)
@@ -831,7 +831,7 @@ def test_get_subnets_no_block_specified(mocker, subtensor):
     mocker.patch.object(subtensor, "query_map_subtensor", return_value=mock_result)
 
     # Call
-    result = subtensor.get_subnets()
+    result = subtensor.get_netuids()
 
     # Asserts
     assert result == [1, 2]
@@ -2196,45 +2196,78 @@ def test_networks_during_connection(mocker):
         sub.chain_endpoint = settings.NETWORK_MAP.get(network)
 
 
-@pytest.mark.parametrize(
-    "fake_value_result",
-    [1, None],
-    ids=["result has value attr", "result has not value attr"],
-)
-def test_get_stake_for_coldkey_and_hotkey(subtensor, mocker, fake_value_result):
-    """Test get_stake_for_coldkey_and_hotkey calls right method with correct arguments."""
+def test_get_stake_for_coldkey_and_hotkey_with_single_result(subtensor, mocker):
+    """Test `get_stake_for_coldkey_and_hotkey` calls right method with correct arguments and get 1 stake info."""
     # Preps
     fake_hotkey_ss58 = "FAKE_H_SS58"
     fake_coldkey_ss58 = "FAKE_C_SS58"
+    fake_netuid = 255
     fake_block = 123
 
-    return_value = (
-        mocker.Mock(value=fake_value_result)
-        if fake_value_result is not None
-        else fake_value_result
+    fake_stake_info_1 = mocker.Mock(hotkey_ss58="some")
+    fake_stake_info_2 = mocker.Mock(
+        hotkey_ss58=fake_hotkey_ss58, netuid=fake_netuid, stake=100
     )
 
-    subtensor.query_subtensor = mocker.patch.object(
-        subtensor, "query_subtensor", return_value=return_value
+    return_value = [
+        fake_stake_info_1,
+        fake_stake_info_2,
+    ]
+
+    subtensor.get_stake_for_coldkey = mocker.patch.object(
+        subtensor, "get_stake_for_coldkey", return_value=return_value
     )
-    spy_balance_from_rao = mocker.spy(subtensor_module.Balance, "from_rao")
 
     # Call
     result = subtensor.get_stake_for_coldkey_and_hotkey(
         hotkey_ss58=fake_hotkey_ss58,
         coldkey_ss58=fake_coldkey_ss58,
+        netuid=fake_netuid,
         block=fake_block,
     )
 
     # Asserts
-    subtensor.query_subtensor.assert_called_once_with(
-        "Stake", fake_block, [fake_hotkey_ss58, fake_coldkey_ss58]
+    subtensor.get_stake_for_coldkey.assert_called_once_with(
+        fake_coldkey_ss58, fake_block
     )
-    if fake_value_result is not None:
-        spy_balance_from_rao.assert_called_once_with(fake_value_result)
-    else:
-        spy_balance_from_rao.assert_not_called()
-    assert result == fake_value_result
+    assert result == fake_stake_info_2
+
+
+def test_get_stake_for_coldkey_and_hotkey_with_multiple_result(subtensor, mocker):
+    """Test `get_stake_for_coldkey_and_hotkey` calls right method with correct arguments and get multiple stake info."""
+    # Preps
+    fake_hotkey_ss58 = "FAKE_H_SS58"
+    fake_coldkey_ss58 = "FAKE_C_SS58"
+    fake_netuid = 255
+    fake_block = 123
+
+    fake_stake_info_1 = mocker.Mock(hotkey_ss58="some")
+    fake_stake_info_2 = mocker.Mock(
+        hotkey_ss58=fake_hotkey_ss58, netuid=fake_netuid, stake=100
+    )
+    fake_stake_info_3 = mocker.Mock(
+        hotkey_ss58=fake_hotkey_ss58, netuid=fake_netuid, stake=200
+    )
+
+    return_value = [fake_stake_info_1, fake_stake_info_2, fake_stake_info_3]
+
+    subtensor.get_stake_for_coldkey = mocker.patch.object(
+        subtensor, "get_stake_for_coldkey", return_value=return_value
+    )
+
+    # Call
+    result = subtensor.get_stake_for_coldkey_and_hotkey(
+        hotkey_ss58=fake_hotkey_ss58,
+        coldkey_ss58=fake_coldkey_ss58,
+        netuid=fake_netuid,
+        block=fake_block,
+    )
+
+    # Asserts
+    subtensor.get_stake_for_coldkey.assert_called_once_with(
+        fake_coldkey_ss58, fake_block
+    )
+    assert result == [fake_stake_info_2, fake_stake_info_3]
 
 
 def test_does_hotkey_exist_true(mocker, subtensor):
@@ -2714,16 +2747,18 @@ def test_add_stake_success(mocker, subtensor):
     fake_wallet = mocker.Mock()
     fake_hotkey_ss58 = "fake_hotkey"
     fake_amount = 10.0
+    fake_netuid = 123
 
     mock_add_stake_extrinsic = mocker.patch.object(
         subtensor_module, "add_stake_extrinsic"
     )
 
     # Call
-    result = subtensor.add_stake(
+    result = subtensor.add_stake_ext(
         wallet=fake_wallet,
         hotkey_ss58=fake_hotkey_ss58,
-        amount=fake_amount,
+        netuid=fake_netuid,
+        tao_amount=fake_amount,
         wait_for_inclusion=True,
         wait_for_finalization=False,
     )
@@ -2733,6 +2768,7 @@ def test_add_stake_success(mocker, subtensor):
         subtensor=subtensor,
         wallet=fake_wallet,
         hotkey_ss58=fake_hotkey_ss58,
+        netuid=fake_netuid,
         amount=fake_amount,
         wait_for_inclusion=True,
         wait_for_finalization=False,
@@ -2746,6 +2782,7 @@ def test_add_stake_multiple_success(mocker, subtensor):
     fake_wallet = mocker.Mock()
     fake_hotkey_ss58 = ["fake_hotkey"]
     fake_amount = [10.0]
+    fake_netuids = [1]
 
     mock_add_stake_multiple_extrinsic = mocker.patch.object(
         subtensor_module, "add_stake_multiple_extrinsic"
@@ -2755,6 +2792,7 @@ def test_add_stake_multiple_success(mocker, subtensor):
     result = subtensor.add_stake_multiple(
         wallet=fake_wallet,
         hotkey_ss58s=fake_hotkey_ss58,
+        netuids=fake_netuids,
         amounts=fake_amount,
         wait_for_inclusion=True,
         wait_for_finalization=False,
@@ -2765,6 +2803,7 @@ def test_add_stake_multiple_success(mocker, subtensor):
         subtensor=subtensor,
         wallet=fake_wallet,
         hotkey_ss58s=fake_hotkey_ss58,
+        netuids=fake_netuids,
         amounts=fake_amount,
         wait_for_inclusion=True,
         wait_for_finalization=False,
@@ -2778,13 +2817,15 @@ def test_unstake_success(mocker, subtensor):
     fake_wallet = mocker.Mock()
     fake_hotkey_ss58 = "hotkey_1"
     fake_amount = 10.0
+    fake_netuid = 123
 
     mock_unstake_extrinsic = mocker.patch.object(subtensor_module, "unstake_extrinsic")
 
     # Call
-    result = subtensor.unstake(
+    result = subtensor.unstake_ext(
         wallet=fake_wallet,
         hotkey_ss58=fake_hotkey_ss58,
+        netuid=fake_netuid,
         amount=fake_amount,
         wait_for_inclusion=True,
         wait_for_finalization=False,
@@ -2795,6 +2836,7 @@ def test_unstake_success(mocker, subtensor):
         subtensor=subtensor,
         wallet=fake_wallet,
         hotkey_ss58=fake_hotkey_ss58,
+        netuid=fake_netuid,
         amount=fake_amount,
         wait_for_inclusion=True,
         wait_for_finalization=False,
@@ -2808,6 +2850,7 @@ def test_unstake_multiple_success(mocker, subtensor):
     fake_wallet = mocker.Mock()
     fake_hotkeys = ["hotkey_1", "hotkey_2"]
     fake_amounts = [10.0, 20.0]
+    fake_netuids = [1, 2]
 
     mock_unstake_multiple_extrinsic = mocker.patch(
         "bittensor.core.subtensor.unstake_multiple_extrinsic", return_value=True
@@ -2817,6 +2860,7 @@ def test_unstake_multiple_success(mocker, subtensor):
     result = subtensor.unstake_multiple(
         wallet=fake_wallet,
         hotkey_ss58s=fake_hotkeys,
+        netuids=fake_netuids,
         amounts=fake_amounts,
         wait_for_inclusion=True,
         wait_for_finalization=False,
@@ -2827,6 +2871,7 @@ def test_unstake_multiple_success(mocker, subtensor):
         subtensor=subtensor,
         wallet=fake_wallet,
         hotkey_ss58s=fake_hotkeys,
+        netuids=fake_netuids,
         amounts=fake_amounts,
         wait_for_inclusion=True,
         wait_for_finalization=False,